requirements.txt
# Core dependencies
ultralytics>=8.1.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
pyyaml>=6.0
tqdm>=4.65.0

# Export dependencies
onnx>=1.15.0
onnxruntime>=1.16.0
onnxsim>=0.4.33

# Optional for TensorRT (if available)
# nvidia-tensorrt>=8.6.0

# Data processing
scikit-image>=0.21.0
matplotlib>=3.7.0
pandas>=2.0.0

# CLI and utilities
argparse
pathlib